requirements.txt
python-telegram-bot
requests
firebase-admin
